app/requirements.txt
Flask==1.1.2
google-cloud-firestore==1.8.1
gunicorn==20.0.4